feat(Like): 상품 조회시 좋아요 반영 + 좋아요 추가/삭제 토글로 변경

+++ src/services/productLike/productLikeService.js
@@ -1,22 +1,22 @@
 import * as productLikeRepository from '@src/repository/productLikeRepository';
-import * as productLikeServiceUtils from '@src/services/productLike/productLikeServiceUtils';
+import { productLikeResonse } from '@src/services/productLike/dto/productLikeResponse';
+import * as productRepository from '@src/repository/productRepository';
+import * as productServieUtils from '@src/services/product/productServiceUtils';
 
-export const addProductLike = async (productId, memberId) => {
-  await productLikeServiceUtils.validateNotExistProductLike(
-    productLikeRepository,
+export const toggleProductLike = async (productId, memberId) => {
+  await productServieUtils.validateExistProduct(productRepository, productId);
+  const findProduct = await productLikeRepository.findProductLike(
     productId,
     memberId
   );
-  await productLikeRepository.save(productId, memberId);
-};
-
-export const deleteProductLike = async (productId, memberId) => {
-  await productLikeServiceUtils.validateExistProductLike(
-    productLikeRepository,
-    productId,
-    memberId
-  );
+  if (findProduct === null) {
+    // 좋아요 되지 않은 경우 좋아요 추가
+    await productLikeRepository.save(productId, memberId);
+    return productLikeResonse(productId, true);
+  }
+  // 좋아요 되있는 경우 좋아요 삭제
   await productLikeRepository.deleteProductLike(productId, memberId);
+  return productLikeResonse(productId, false);
 };
 
 export const checkIsProductLike = async (productId, memberId) => {
